+++ demo/cypress/DropDown/index.tsx
@@ -3,13 +3,15 @@ import * as React from 'react';
 import * as L from '../../../korus-ui';
 
 export const DropDown = (): React.ReactElement => {
+  const [isDisabled, setIsDisabled] = React.useState(false);
 
   return (
     <>
       <L.P>interactionMode default</L.P>
-      <L.Div _demoStory _flexRow>
+      <L.Div _demoStory>
         <L.DropDown
-          wrapperRender={({elementProps}: any) => <L.Button {...elementProps}/>}
+          isDisabled={isDisabled}
+          wrapperRender={({ elementProps }: L.commonTypes.RenderEvent<L.DropDownTypes.DropDownProps>): React.ReactNode => <L.Button {...elementProps} />}
           id="DDButtonHover"
         >
           <L.Span>Hover over me</L.Span>
@@ -28,8 +30,12 @@ export const DropDown = (): React.ReactElement => {
             </L.Li>
           </L.Ul>
         </L.DropDown>
-        
+
+        <br />
+        <br />
+
         <L.DropDown
+          isDisabled={isDisabled}
           id="DDDivHover"
         >
           <L.Span>Hover over me</L.Span>
@@ -55,9 +61,10 @@ export const DropDown = (): React.ReactElement => {
       <br />
       <br />
       <L.P>interactionMode click</L.P>
-      <L.Div _demoStory _flexRow>   
+      <L.Div _demoStory>
         <L.DropDown
-          wrapperRender={({elementProps}: any) => <L.Button {...elementProps}/>}
+          isDisabled={isDisabled}
+          wrapperRender={({ elementProps }: L.commonTypes.RenderEvent<L.DropDownTypes.DropDownProps>): React.ReactNode => <L.Button {...elementProps} />}
           interactionMode="click"
           id="DDButtonClick"
         >
@@ -77,8 +84,12 @@ export const DropDown = (): React.ReactElement => {
             </L.Li>
           </L.Ul>
         </L.DropDown>
-        
+
+        <br />
+        <br />
+
         <L.DropDown
+          isDisabled={isDisabled}
           interactionMode="click"
           id="DDDivClick"
         >
@@ -99,6 +110,16 @@ export const DropDown = (): React.ReactElement => {
           </L.Ul>
         </L.DropDown>
       </L.Div>
+
+      <br />
+      <br />
+
+      <L.Button
+        id="disableButton"
+        _warning
+        onClick={(): void => setIsDisabled(!isDisabled)}
+      >Disable/Enable
+      </L.Button>
     </>
-  )
+  );
 };